pyproject.toml
[project]
name = "catalog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "fire>=0.7.1",
    "numpy>=2.3.4",
    "openai>=1.109.1",
    "pgai[semantic-catalog]>=0.12.1",
    "psycopg>=3.2.12",
    "psycopg-pool>=3.2.7",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

[tool.uv]
package = true

[tool.setuptools]
py-modules = ["main"]

[project.scripts]
catalog = "main:main"
